add: send-otp

+++ NET1806_LittleJoyStore/NET1806_LittleJoy.Repository/Entities/OrderDetail.cs
@@ -11,7 +11,7 @@ public partial class OrderDetail
 
     public int? ProductId { get; set; }
 
-    public double? Price { get; set; }
+    public int? Price { get; set; }
 
     public int? Quantity { get; set; }
